Move user logged in guard to general users folder
