+++ ClientApp/KachnaOnline/src/app/users/user-logged-in.guard.ts
@@ -4,7 +4,7 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
-import { AuthenticationService } from "../../shared/services/authentication.service";
+import { AuthenticationService } from "../shared/services/authentication.service";
 
 @Injectable({
   providedIn: 'root'
